Renamed the 'scripts' package to 'ria_cli'. This was done so the package is easier to identify within site-packages
